CASMCMS-2090: Wait for Gitea API to be available and ready
Wait for the gitea API to be available before running the import
script, and wait for the API to return a successful status code
before continuing on with the import. v1.2.4 was changed to wait
for Gitea to be available, but not to return a successful status.
This allows for the import script to attempt to create Gitea
resources before the Gitea API is actually able to accept them.