CASMINST-3525: remove branch protections instead of enabling push
Previously, when a branch already existed the was being pushed,
the import.py script used the Gitea API to enable pushes to the
repository after branches had been protected from a previous import.
When the branch does not change, ie a new version of CSM, but not a
new version of csm-import, a force push happens, but branch protections
that allow pushing on a protected branch do not allow force pushes.
To remedy this, instead of just enabling push, remove the branch
protections altogether, force push the branch, and then re-enable
branch protections if requested.