CASMCMS-7890: Introduce CF_IMPORT_FORCE_EXISTING_BRANCH parameter
In order to allow for repeated calls to cf-gitea-import with the
same content, which may happen during some product upgrades, do
not import the content again so the branch and its commit id are
preserved. Previously, cf-gitea-import would reset the branch and
create a new commit no matter what.

If repeated attempts to upload a version are made when the branch
already exists, and the user wishes to create a new branch with a
new commit id, the CF_IMPORT_FORCE_EXISTING_BRANCH parameter can
force cf-gitea-import to do so.

Also added is a new way to start up a local gitea instance in a
container to test cf-gitea-import in isolation.

+semver: minor