CRAYSAT-1512: Fix incorrect use of urljoin
This commit fixes the use of `urljoin` so that trailing path components
are not stripped off of `gitea_base_url`.

If the second argument to `urljoin()` begins with a `/` character
it is assumed that it begins at the root of the website and trailing
path components are stripped from the base URL (the first argument).
For example, `urljoin('https://foo.com/bar', '/baz') returns:
`'https://foo.com/baz'`.

Additionally, if the first argument to `urljoin()` contains path
components after the domain and does NOT end with a `/` character
(e.g. https://foo.com/bar), the path components are stripped.
For example, `urljoin('https://foo.com/bar', 'baz')` returns
`'https://foo.com/baz'` as well.

This usually isn't encountered because the typical way of using
cf-gitea-import is using `'https://gitea-vcs'` as the value of
`gitea_base_url`.

This commit also removes an unused import of HTTPBasicAuth in `import.py`,
as well as the now-unused import of urljoin.

Finally, add a change log entry.

Test Description:
* Built the image.
* Built a version of sat-cfs-install derived from this image, and used
  a loftsman manifest and helm chart to import a version of SAT
  configuration.
* Built a version of the product stream with this image, Ran cf-gitea-import
  directly using podman, mounting in a version of SAT configuration
  over `/content`.