CRAYSAT-1512: Add alternate entrypoint
This commit adds an entrypoint that can be used to run the
cf-gitea-import image directly (e.g. via podman) instead of using
Kubernetes and Helm. The motivation for this is so that products
can mount the configuration in the `/content` directory rather than
create an image based on cf-gitea-import.

Add change log entry.

Test Description:
* Built the image.
* Built a version of sat-cfs-install derived from this image, and used
  a loftsman manifest and helm chart to import a version of SAT
  configuration.
* Built a version of the product stream with this image, then ran
  cf-gitea-import directly using podman, mounting in a version of
  SAT configuration over `/content`.